refactor ♻️: Refactor CLI structure for better separation of concerns
This commit refactors the command-line interface (CLI) structure in `git_acp/cli` to improve code organization and maintainability. The changes include renaming `main.py` to `cli.py`, updating imports, and enhancing the documentation. This refactoring enhances the readability and maintainability of the CLI, making it easier for developers to understand and modify the command-line interface in the future.
